+++ fractal/components/assessments/AssessmentScore/index.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+
+import AssessmentScore from './AssessmentScore.jsx';
+import config from './AssessmentScore.config';
+
+ReactDOM.render(
+    <AssessmentScore {...config.context} />,
+    document.getElementsByClassName('base-preview')[0]
+);